multi_agent_whitepaper/requirements.txt
openai>=0.29.0

python-dotenv
